Correct `ETS_TOOLKIT` for current mayavi release
The current stable release of `mayavi` (published October 2022) relies on an earlier version of the `traitsui` API. Enthought's suggestion for code that needs to use this is to set the `ETS_TOOLKIT` (Enthought Tool Suite) to `qt4`. This adds some backwards-compatibility code that maps the legacy imports to the updated imports.

Once Enthought publishes a new release of `mayavi`,, all uses of `os.environ['ETS_TOOLKIT'] = 'qt4'` will probably need to be removed or replaced.